Firestoreのコードの管理の仕方を変更する
