+++ app/src/main/java/jp/co/cyberagent/dojo2020/DI.kt
@@ -7,8 +7,8 @@ import jp.co.cyberagent.dojo2020.data.DefaultMemoRepository
 import jp.co.cyberagent.dojo2020.data.MemoRepository
 import jp.co.cyberagent.dojo2020.data.local.LocalMemoDataSource
 import jp.co.cyberagent.dojo2020.data.local.db.MemoDataBase
+import jp.co.cyberagent.dojo2020.data.remote.firestore.DefaultFireStoreDataSource
 import jp.co.cyberagent.dojo2020.data.remote.firestore.FireStoreDataSource
-import jp.co.cyberagent.dojo2020.data.remote.RemoteMemoDataSource
 
 object DI {
     fun injectRepository(context: Context): MemoRepository {
@@ -25,7 +25,7 @@ object DI {
     }
 
     private fun injectRemoteDataSource(): FireStoreDataSource {
-        return RemoteMemoDataSource()
+        return DefaultFireStoreDataSource()
     }
 
     private fun injectDatabase(context: Context): MemoDataBase {
